Move common classes into common dir
